+++ src/click/shell_completion.py
@@ -167,25 +167,25 @@ fi
 """
 
 _SOURCE_FISH = """\
-function %(complete_func)s
+function %(complete_func)s;
     set -l response (env %(complete_var)s=fish_complete COMP_WORDS=(commandline -cp) \
-COMP_CWORD=(commandline -t) %(prog_name)s)
+COMP_CWORD=(commandline -t) %(prog_name)s);
 
-    for completion in $response
-        set -l metadata (string split "," $completion)
+    for completion in $response;
+        set -l metadata (string split "," $completion);
 
-        if test $metadata[1] = "dir"
-            __fish_complete_directories $metadata[2]
-        else if test $metadata[1] = "file"
-            __fish_complete_path $metadata[2]
-        else if test $metadata[1] = "plain"
-            echo $metadata[2]
-        end
-    end
-end
+        if test $metadata[1] = "dir";
+            __fish_complete_directories $metadata[2];
+        else if test $metadata[1] = "file";
+            __fish_complete_path $metadata[2];
+        else if test $metadata[1] = "plain";
+            echo $metadata[2];
+        end;
+    end;
+end;
 
 complete --no-files --command %(prog_name)s --arguments \
-"(%(complete_func)s)"
+"(%(complete_func)s)";
 """